ref(api) Move REST API project key serializer to dedicated module
