+++ src/sentry/api/endpoints/project_key_details.py
@@ -1,21 +1,18 @@
 from __future__ import absolute_import
 
 from django.db.models import F
-from rest_framework import serializers, status
+from rest_framework import status
 from rest_framework.response import Response
 
 from sentry import features
 from sentry.api.base import DocSection
 from sentry.api.bases.project import ProjectEndpoint
 from sentry.api.exceptions import ResourceDoesNotExist
-from sentry.api.fields.empty_integer import EmptyIntegerField
 from sentry.api.serializers import serialize
+from sentry.api.serializers.rest_framework import ProjectKeySerializer
 from sentry.models import AuditLogEntryEvent, ProjectKey, ProjectKeyStatus
 from sentry.utils.apidocs import scenario, attach_scenarios
-from sentry.loader.browsersdkversion import (
-    get_default_sdk_version_for_project,
-    get_browser_sdk_version_choices,
-)
+from sentry.loader.browsersdkversion import get_default_sdk_version_for_project
 
 
 @scenario("DeleteClientKey")
@@ -36,20 +33,6 @@ def update_key_scenario(runner):
         path="/projects/%s/%s/keys/%s/"
         % (runner.org.slug, runner.default_project.slug, key.public_key),
         data={"name": "Quite Positive Key"},
-    )
-
-
-class RateLimitSerializer(serializers.Serializer):
-    count = EmptyIntegerField(min_value=0, required=False, allow_null=True)
-    window = EmptyIntegerField(min_value=0, max_value=60 * 60 * 24, required=False, allow_null=True)
-
-
-class KeySerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=200, required=False, allow_blank=True, allow_null=True)
-    isActive = serializers.BooleanField(required=False)
-    rateLimit = RateLimitSerializer(allow_null=True)
-    browserSdkVersion = serializers.ChoiceField(
-        choices=get_browser_sdk_version_choices(), required=False
     )
 
 
@@ -88,7 +71,7 @@ class ProjectKeyDetailsEndpoint(ProjectEndpoint):
         except ProjectKey.DoesNotExist:
             raise ResourceDoesNotExist
 
-        serializer = KeySerializer(data=request.data, partial=True)
+        serializer = ProjectKeySerializer(data=request.data, partial=True)
         default_version = get_default_sdk_version_for_project(project)
 
         if serializer.is_valid():
