+++ src/sentry/api/endpoints/project_keys.py
@@ -1,14 +1,14 @@
 from __future__ import absolute_import
 
 from django.db.models import F
-from rest_framework import serializers, status
+from rest_framework import status
 from rest_framework.response import Response
 
 from sentry import features
 from sentry.api.base import DocSection
 from sentry.api.bases.project import ProjectEndpoint
-from sentry.api.fields.empty_integer import EmptyIntegerField
 from sentry.api.serializers import serialize
+from sentry.api.serializers.rest_framework import ProjectKeySerializer
 from sentry.models import AuditLogEntryEvent, ProjectKey, ProjectKeyStatus
 from sentry.utils.apidocs import scenario, attach_scenarios
 
@@ -27,18 +27,6 @@ def create_key_scenario(runner):
         path="/projects/%s/%s/keys/" % (runner.org.slug, runner.default_project.slug),
         data={"name": "Fabulous Key"},
     )
-
-
-class RateLimitSerializer(serializers.Serializer):
-    count = EmptyIntegerField(min_value=0, required=False, allow_null=True)
-    window = EmptyIntegerField(min_value=0, max_value=60 * 60 * 24, required=False, allow_null=True)
-
-
-class KeySerializer(serializers.Serializer):
-    name = serializers.CharField(max_length=64, required=False, allow_blank=True, allow_null=True)
-    public = serializers.RegexField(r"^[a-f0-9]{32}$", required=False, allow_null=True)
-    secret = serializers.RegexField(r"^[a-f0-9]{32}$", required=False, allow_null=True)
-    rateLimit = RateLimitSerializer(required=False, allow_null=True)
 
 
 class ProjectKeysEndpoint(ProjectEndpoint):
@@ -90,7 +78,7 @@ class ProjectKeysEndpoint(ProjectEndpoint):
                                      belong to.
         :param string name: the name for the new key.
         """
-        serializer = KeySerializer(data=request.data)
+        serializer = ProjectKeySerializer(data=request.data)
 
         if serializer.is_valid():
             result = serializer.validated_data
